add rectangle
