+++ atest/API/04_Shapes_GRAPH.robot.j2
@@ -1,6 +1,7 @@
 {% set shapes = {
     "Text": ["fill", "background", "stroke"],
     "Ellipse": ["fill", "stroke"],
+    "Rectangle": ["fill", "stroke"],
 } %}
 *** Settings ***
 Resource        ../_resources/keywords/Server.robot
