+++ js/widgets/behaviors/base.ts
@@ -3,9 +3,7 @@
  * Distributed under the terms of the Modified BSD License.
  */
 import type { Template } from 'nunjucks';
-import type THREE from 'three';
 
-import { JSONExt } from '@lumino/coreutils';
 import { ISignal, Signal } from '@lumino/signaling';
 
 import { IBackboneModelOptions, WidgetModel } from '@jupyter-widgets/base';
@@ -18,13 +16,10 @@ import {
   IHasGraph,
   ILinkBehaveOptions,
   INodeBehaveOptions,
-  INodeCanvasBehaveOptions,
-  INodeThreeBehaveOptions,
   TUpdateKind,
   WIDGET_DEFAULTS,
 } from '../../tokens';
 import { getCoercer, noop } from '../../utils';
-import { functor } from '../../utils';
 
 export class BehaviorModel extends WidgetModel implements IBehave {
   protected _updateRequested: Signal<IBehave, TUpdateKind>;
@@ -214,109 +209,5 @@ export class LinkColumnOrTemplateModel
     }
 
     return value || null;
-  }
-}
-
-export class ShapeBaseModel extends BehaviorModel {
-  /** Required in subclass. The model name should be unique between shapes.  */
-  static model_name = 'TextShapeModel';
-
-  /** Required in subclass. All new traits of a shape might be dynamic  */
-  static serializers = {
-    ...BehaviorModel.serializers,
-  };
-
-  /** Required in subclass. Provides acesss to active `__class__`. */
-  protected get _modelClass(): typeof ShapeBaseModel {
-    return ShapeBaseModel;
-  }
-
-  /** Required in subclass. Draw a shape on a canvas. */
-  drawNode2D(options: INodeCanvasBehaveOptions): void {
-    return;
-  }
-
-  /** Required in subclass. Draw a shape in Three.js. */
-  drawNode3D(options: INodeThreeBehaveOptions): THREE.Object3D | null {
-    return;
-  }
-
-  /** Facets are cached as handlers for a specific entity. */
-  protected _facets: Record<string, Function> = JSONExt.emptyObject as any;
-
-  /** Names of facets are calculated once, on initialization. */
-  protected _facetNames: string[] | null = null;
-
-  /** Lazily calculate asset names. */
-  protected get facetNames() {
-    if (this._facetNames == null) {
-      const baseKeys = [...Object.keys(ShapeBaseModel.serializers)];
-      const facetNames: string[] = [];
-      for (const key of Object.keys(this._modelClass.serializers)) {
-        if (baseKeys.includes(key)) {
-          continue;
-        }
-        facetNames.push(key);
-      }
-      this._facetNames = facetNames;
-    }
-    return this._facetNames;
-  }
-
-  /** Initialize the model and wire up listeners.  */
-  initialize(attributes: Backbone.ObjectHash, options: IBackboneModelOptions) {
-    super.initialize(attributes, options);
-    let events = '';
-    for (const facet of this.facetNames) {
-      events += `change:${facet} `;
-    }
-    this.on(events, this._onFacetsChanged, this);
-    void this._onFacetsChanged.call(this);
-  }
-
-  /** Populate facet handlers. */
-  async ensureFacets() {
-    const facets: Record<string, Function> = {};
-    for (const facetName of this.facetNames) {
-      let facet = this.get(facetName);
-      if (facet instanceof DynamicModel) {
-        await facet.ensureHandlers();
-        facets[facetName] = facet.nodeHandler;
-        facet.updateRequested.connect(this._onFacetsChanged, this);
-        continue;
-      }
-
-      if (facet != null) {
-        facets[facetName] = functor(facet);
-      }
-    }
-    this._facets = facets;
-  }
-
-  /** Evaluate all facets with the runtime shape*/
-  protected _resolveFacets(
-    options: INodeCanvasBehaveOptions | INodeThreeBehaveOptions
-  ): Record<string, any> {
-    const draw: Record<string, any> = {};
-    for (const facetName of this._facetNames) {
-      if (this._facets[facetName]) {
-        try {
-          draw[facetName] = this._facets[facetName](options);
-        } catch (err) {
-          console.warn(`${EMOJI} encountered error for ${facetName}`, options, err);
-        }
-      }
-    }
-    return draw;
-  }
-
-  /** Handle the fact changing. */
-  protected async _onFacetsChanged() {
-    this._facets = JSONExt.emptyObject as any;
-    this._updateRequested.emit(void 0);
-  }
-
-  defaults() {
-    return { ...super.defaults(), _model_name: this._modelClass.model_name };
   }
 }
